rename Sidebar to ConversationList, better understanding, refactor code to have all that concerns this component in the same place
