+++ examples/server/webui/src/App.tsx
@@ -1,6 +1,6 @@
 import { HashRouter, Outlet, Route, Routes } from 'react-router';
 import Header from './components/Header';
-import Sidebar from './components/Sidebar';
+import ConversationList from './components/ConversationList';
 import { AppContextProvider, useAppContext } from './utils/app.context';
 import ChatScreen from './components/ChatScreen';
 import SettingDialog from './components/SettingDialog';
@@ -26,7 +26,7 @@ function AppLayout() {
   const { showSettings, setShowSettings, settingsSeed } = useAppContext();
   return (
     <>
-      <Sidebar />
+      <ConversationList />
       <div
         className="drawer-content grow flex flex-col h-screen w-screen mx-auto px-4 overflow-auto"
         id="main-scroll"
